+++ apps/bot/src/events/event-handlers.ts
@@ -1,7 +1,7 @@
 import { ReadyHandler } from './ready/ready.handler';
 import { WarnHandler } from './warn/warn.handler';
 import { DiscordMemberJoinedHandler } from './discord-member-joined/discord-member-joined.handler';
-import { DiscordMemberRolesUpdatedHandler } from './discord-member-updated/discord-member-roles-updated.handler';
+import { DiscordMemberRolesUpdatedHandler } from './discord-member-roles-updated/discord-member-roles-updated.handler';
 
 export const EVENT_HANDLERS = [
   ReadyHandler,
